Fix -Wconversion/-Wsign-conversion warnings in 13 modules
Make integer conversions explicit with static_cast at the point of
conversion across the db, fastx I/O, attributes, orient, otutable and
related modules. Conversions are non-negative lengths, counts, indices,
abundances, quality symbols (indexed through unsigned char), and
pointer differences; existing 32-bit mask truncations and the int
length parameters of the print/header APIs are preserved verbatim.
No behavior change.

+++ src/arch.cc
@@ -90,7 +90,7 @@ auto arch_get_memused() -> uint64_t
   return r_usage.ru_maxrss;
 # else
   /* Linux: ru_maxrss gives the size in kilobytes  */
-  return r_usage.ru_maxrss * 1024;
+  return static_cast<uint64_t>(r_usage.ru_maxrss) * 1024;
 # endif
 
 #endif
@@ -134,7 +134,7 @@ auto arch_get_memtotal() -> uint64_t
     {
       fatal("Cannot determine amount of RAM");
     }
-  return pagesize * phys_pages;
+  return static_cast<uint64_t>(pagesize * phys_pages);
 
 #else
 
@@ -177,10 +177,10 @@ auto arch_get_user_system_time(double * user_time, double * system_time) -> void
 #else
   struct rusage r_usage;
   getrusage(RUSAGE_SELF, & r_usage);
-  * user_time = (r_usage.ru_utime.tv_sec * 1.0)
-    + (r_usage.ru_utime.tv_usec * 1.0e-6);
-  * system_time = (r_usage.ru_stime.tv_sec * 1.0)
-    + (r_usage.ru_stime.tv_usec * 1.0e-6);
+  * user_time = (static_cast<double>(r_usage.ru_utime.tv_sec) * 1.0)
+    + (static_cast<double>(r_usage.ru_utime.tv_usec) * 1.0e-6);
+  * system_time = (static_cast<double>(r_usage.ru_stime.tv_sec) * 1.0)
+    + (static_cast<double>(r_usage.ru_stime.tv_usec) * 1.0e-6);
 #endif
 }
 
@@ -265,7 +265,7 @@ auto xlseek(int file_descriptor, uint64_t offset, int whence) -> uint64_t
 #ifdef _WIN32
   return _lseeki64(file_descriptor, offset, whence);
 #else
-  return lseek(file_descriptor, offset, whence);  // libC or linuxism: replace with std::fseek()?
+  return static_cast<uint64_t>(lseek(file_descriptor, static_cast<off_t>(offset), whence));  // libC or linuxism: replace with std::fseek()?
 #endif
 }
 
@@ -276,7 +276,7 @@ auto xftello(std::FILE * stream) -> uint64_t
 #ifdef _WIN32
   return _ftelli64(stream);
 #else
-  return ftello(stream);
+  return static_cast<uint64_t>(ftello(stream));
 #endif
 }
